Libs/ExtendedSeleniumLibrary/requirements.txt
# This file contains all the necessary dependencies for Extended Library

robotframework-seleniumlibrary==4.1.0
